# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoyaStar Backend API",
        "description": "LMS platform for managing students, teachers, schools, and curriculum.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/cities": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "List all cities",
                "operationId": "e25de57040f79e34bde45fbad1555991",
                "parameters": [
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Set to true to get all cities without pagination",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort column (e.g. name, created_at)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_type",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default 20)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cities"
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/admin/cities": {
            "post": {
                "tags": [
                    "Cities"
                ],
                "summary": "Create city",
                "operationId": "ca44ee4f9aa208894481777e393d5347",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "properties": {
                                            "az": {
                                                "type": "string"
                                            },
                                            "en": {
                                                "type": "string"
                                            },
                                            "ru": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "City created"
                    }
                }
            }
        },
        "/cities/{city}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get city by ID",
                "operationId": "28e843404afd450f4e22680c1c42c0ba",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "City ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City data"
                    }
                }
            }
        },
        "/admin/cities/{city}": {
            "put": {
                "tags": [
                    "Cities"
                ],
                "summary": "Update city",
                "operationId": "a87c8e245f7a73fc79b01cf7e0517cb8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "properties": {
                                            "az": {
                                                "type": "string"
                                            },
                                            "en": {
                                                "type": "string"
                                            },
                                            "ru": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "City updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cities"
                ],
                "summary": "Delete city",
                "operationId": "50aa46271b0540390d7aca973e963239",
                "responses": {
                    "200": {
                        "description": "City deleted"
                    }
                }
            }
        },
        "/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "List all grades",
                "operationId": "76fce6f015ff6699e282c0374f2029a4",
                "parameters": [
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Set to true to get all grades without pagination",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort column (e.g. name, created_at)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_type",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default 20)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of grades"
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/admin/grades": {
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Create grade",
                "operationId": "b16174552b1284b1209bc0012c743b7f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Grade created"
                    }
                }
            }
        },
        "/admin/grades/{grade}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Get grade by ID",
                "operationId": "d0a0cec83ca83cd8f9b164aacccbfac6",
                "parameters": [
                    {
                        "name": "grade",
                        "in": "path",
                        "description": "Grade ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grade data"
                    }
                }
            },
            "put": {
                "tags": [
                    "Grades"
                ],
                "summary": "Update grade",
                "operationId": "fd13d206ecada4682e8e1fbba33c5fbd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grade updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grades"
                ],
                "summary": "Delete grade",
                "operationId": "e7d830c91b854114034044d74abf5261",
                "responses": {
                    "200": {
                        "description": "Grade deleted"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "bearer",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}